WebExceptionHandler reactivo añadido

+++ src/main/java/com/app/service/impl/CRUDImpl.java
@@ -18,7 +18,8 @@ public abstract class CRUDImpl<T,ID> implements ICRUD<T,ID> {
 
     @Override
     public Mono<T> update(ID id, T t) {
-        return getRepo().save(t);
+        return getRepo().findById(id)
+                .flatMap(e->getRepo().save(t));
     }
 
     @Override
@@ -33,6 +34,14 @@ public abstract class CRUDImpl<T,ID> implements ICRUD<T,ID> {
 
     @Override
     public Mono<Boolean> delete(ID id) {
-        return getRepo().deleteById(id).thenReturn(true);
+        return getRepo().findById(id)
+                .hasElement()
+                .flatMap(result->{
+                    if(result){
+                        return getRepo().deleteById(id).thenReturn(true);
+                    }else {
+                        return Mono.just(false);
+                    }
+                });
     }
 }
